fix: add PeopleTable component and move from a to Link
