--- src/components/Home/Home.tsx
@@ -1,7 +1,0 @@
-export const Home = () => {
-  return (
-    <>
-      <h1 className="title">Home Page</h1>
-    </>
-  );
-};
