--- src/components/NotFound/NotFound.tsx
@@ -1,7 +1,0 @@
-export const NotFound = () => {
-  return (
-    <>
-      <h1 className="title">Page not found</h1>
-    </>
-  );
-};
