+++ src/components/People/index.ts
